fix : 객체 변경에 따른 컴파일 에러 해결
