+++ src/main/java/com/book_everywhere/domain/tagged/TaggedRepository.java
@@ -14,27 +14,16 @@ import java.util.List;
 @Repository
 public interface TaggedRepository extends JpaRepository<Tagged,Long> {
     // 선택한 태그에 1번이라도 Tag당한 핀 모두 호출
-    @Query("SELECT pin FROM Pin pin JOIN pin.tags tagged WHERE tagged.tag.content = :content")
-    List<Pin> mFindTaggedPin(@Param("content") String content);
-
-
     @Query("SELECT t FROM Tagged t WHERE t.tag.id = :tagId AND t.pin.id = :pinId")
     Tagged mFindTagged(@Param("tagId") Long tagId, @Param("pinId") Long pinId);
 
-    //pin에 대한 모든 태그 개수 새기
-// Tagged 엔티티의 Repository 인터페이스 내에 메소드 추가
-    Page<Tagged> findByPinIdOrderByCountDesc(Long pinId, Pageable pageable);
-
-    @Query("SELECT t FROM Tagged t WHERE t.pin.id = :pinId ORDER BY t.count DESC")
-    List<Tagged> mFindPinWithTagCount(@Param("pinId") Long pinId, Pageable pageable);
-
-    @Modifying
-    @Query("DELETE FROM Tagged WHERE id = :taggedId")
-    int mDeleteTagged(@Param("taggedId") Long taggedId);
+    //pin에 tag들에 대한 count개수 상위 5개
+    @Query("SELECT t.tag.content, COUNT(tagged) FROM tagged t  WHERE t.pin.id = :pinId GROUP BY t.tag.content ORDER BY count DESC")
+    Page<Object[]> mCountByPinId(@Param("pinId") Long pinId, Pageable pageable);
 
     //핀에 해당하는 모든 태그 조회하기 - WebDataService에서 사용하는 용도
      List<Tagged> findAllByPinId(Long pinId);
 
-     @Query("SELECT t FROM Tagged t JOIN t.pin.visits v WHERE t.pin.id = :pinId AND v.user.socialId = :socialId")
+     @Query("SELECT t FROM Tagged t WHERE t.pin.id = :pinId AND t.user.socialId = :socialId")
      List<Tagged> mFindAllByPinAndUser(@Param("pinId") Long pinId, @Param("socialId") Long socialId);
 }
