+++ src/main/java/com/book_everywhere/web/TagController.java
@@ -1,11 +1,9 @@
 package com.book_everywhere.web;
 
-import com.book_everywhere.service.TagService;
+import com.book_everywhere.service.TaggedService;
 import com.book_everywhere.web.dto.CMRespDto;
 import com.book_everywhere.web.dto.tag.TagDto;
-import com.book_everywhere.web.dto.tag.TaggedDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,17 +14,10 @@ import java.util.List;
 @RequiredArgsConstructor
 public class TagController {
 
-    private final TagService tagService;
+    private final TaggedService taggedService;
     @GetMapping("/api/tags")
     public CMRespDto<?> findAllTag() {
-        List<TagDto> result = tagService.모든태그조회();
+        List<TagDto> result = taggedService.모든태그조회();
         return new CMRespDto<>(HttpStatus.OK, result, "태그 조회 성공");
     }
-
-    @GetMapping("/api/tags/top")
-    public CMRespDto<?> findTagCount() {
-        List<Page<TaggedDto>> result = tagService.핀의5개태그조회();
-        return new CMRespDto<>(HttpStatus.OK, result, "태그 개수 조회 성공");
-    }
-
 }
